`SequencerDataWrapper` does not transform list to dict on validation
Instead that functionality has moved to method. The problem with transforming on
validation is that if the model is validated once, it cannot be dumped to
`dict`/`json` and then parsed back to a model again. This way the model can be
transformed immediately before the change of shape is needed.